2 commit

+++ app/src/main/java/com/example/android_ejercicios_xml/domain/models/Cronograma.kt
@@ -2,7 +2,7 @@ package com.example.android_ejercicios_xml.domain.models
 
 data class Cronograma(
     val vencimiento: String,
-    val amortizacion: Double,
+    val amortizacion: Float,
     val interes: Double,
     val seguros: Double,
     val subvencion: Double,
